Fix warnings and `R_ASSERT` logic

+++ src/3rd party/crypto/openssl/src/a_d2i_fp.c
@@ -201,7 +201,7 @@ static int asn1_d2i_read_bio(BIO *in, BUF_MEM **pb)
 			else
 				ERR_clear_error(); /* clear error */
 			}
-		i=c.p-p;/* header length */
+		i= (int)(c.p-p);/* header length */
 		off+=i;	/* end of data */
 
 		if (c.inf & 1)
