+++ src/Layers/xrRender/r_constants.cpp
@@ -261,7 +261,7 @@ void R_constant_table::clear	()
 BOOL R_constant_table::equal(R_constant_table& C)
 {
 	if (table.size() != C.table.size())	return FALSE;
-	u32 size			= table.size();
+	u32 size			= (u32)table.size();
 	for (u32 it=0; it<size; it++)
 	{
 		if (!table[it]->equal(&*C.table[it]))	return FALSE;
