+++ src/xrEngine/XR_IOConsole_control.cpp
@@ -29,7 +29,7 @@ void CConsole::next_cmd_history_idx() {
 void CConsole::prev_cmd_history_idx() {
 	++m_cmd_history_idx;
 	if (m_cmd_history_idx >= (int)m_cmd_history.size()) {
-		m_cmd_history_idx = m_cmd_history.size() - 1;
+		m_cmd_history_idx = (u32)m_cmd_history.size() - 1;
 	}
 }
 
@@ -44,7 +44,7 @@ void CConsole::next_selected_tip() {
 
 void CConsole::check_next_selected_tip() {
 	if (m_select_tip >= (int)m_tips.size()) {
-		m_select_tip = m_tips.size() - 1;
+		m_select_tip = (u32)m_tips.size() - 1;
 	}
 
 	int sel_dif = m_select_tip - VIEW_TIPS_COUNT + 1;
