+++ src/xrSound/SoundRender_TargetA.cpp
@@ -201,13 +201,13 @@ void	CSoundRender_TargetA::fill_block(ALuint BufferID)
 
     if (format == AL_FORMAT_MONO16)
     {
-        m_pEmitter->fill_block(&g_target_temp_data.front(), g_target_temp_data.size());
-        A_CHK(alBufferData(BufferID, format, &g_target_temp_data.front(), g_target_temp_data.size(), m_pEmitter->source()->m_wformat.nSamplesPerSec));
+        m_pEmitter->fill_block(&g_target_temp_data.front(), (u32)g_target_temp_data.size());
+        A_CHK(alBufferData(BufferID, format, &g_target_temp_data.front(), (int)g_target_temp_data.size(), m_pEmitter->source()->m_wformat.nSamplesPerSec));
     }
     else
     {
-        m_pEmitter->fill_block(&g_target_temp_data_16.front(), g_target_temp_data.size());
-        A_CHK(alBufferData(BufferID, format, &g_target_temp_data_16.front(), g_target_temp_data.size(), m_pEmitter->source()->m_wformat.nSamplesPerSec));
+        m_pEmitter->fill_block(&g_target_temp_data_16.front(), (u32)g_target_temp_data.size());
+        A_CHK(alBufferData(BufferID, format, &g_target_temp_data_16.front(), (int)g_target_temp_data.size(), m_pEmitter->source()->m_wformat.nSamplesPerSec));
     }
 }
